requirement.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.44.0
django-crispy-forms==1.7.2
django-registration-redux==2.8
djongo==1.3.3
idna==2.10
mysqlclient==2.0.2
oauthlib==3.1.0
Pillow==6.2.0
pycparser==2.20
PyJWT==2.0.0
pymongo==3.11.2
python3-openid==3.2.0
pytz==2019.2
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.2
